Programando inicio de sesión y registro de nuevos maestros

+++ assets/js/panel.js
@@ -5,22 +5,39 @@ var tableTecnicas = document.getElementById("tableBodyTecnicas");
 var alumnoDiplomados = "";
 var alumnoTecnicas = "";
 
+$("#btnModal").css('cursor', 'pointer');
+$("#logOut").css('cursor', 'pointer');
+$(".txtRojo").hide();
+$(".txtVerde").hide();
+
+
 window.addEventListener('DOMContentLoaded', async(e) => {
-    /* Traemos el arreglo de alumnos */
-    const querySnapshot = await db.collection('alumnos').get();
 
-    // Consultamos y llenamos la tabla
-    consultarAlumnos(querySnapshot);
+    auth.onAuthStateChanged(async user => {
+        console.log("Hubo en cambio en el Auth");
+        tableDiplomados.innerHTML = '';
+        tableTecnicas.innerHTML = '';
+        if (user) {
+            /* Traemos el arreglo de alumnos */
+            const querySnapshot = await db.collection('alumnos').get();
 
-    // Eliminar a un alumno
-    const btnEliminar = document.querySelectorAll(".btnDelete");
-    btnEliminar.forEach(btn => {
-        btn.addEventListener('click', async(e) => {
-            id = e.target.dataset.id;
-            await db.collection('alumnos').doc(id).delete();
-            location.reload();
-        });
+            // Consultamos y llenamos la tabla
+            consultarAlumnos(querySnapshot);
+
+            // Eliminar a un alumno
+            const btnEliminar = document.querySelectorAll(".btnDelete");
+            btnEliminar.forEach(btn => {
+                btn.addEventListener('click', async(e) => {
+                    id = e.target.dataset.id;
+                    await db.collection('alumnos').doc(id).delete();
+                    location.reload();
+                });
+            });
+        } else {
+            location.href = "admin.html";
+        }
     });
+
 });
 
 function consultarAlumnos(querySnapshot) {
@@ -133,4 +150,109 @@ function definirCurso(numCurso) {
     }
 
     return curso;
+}
+
+// Registrando a un maestro
+
+var emailCorrect = false;
+var passwordCorrect = false;
+
+
+$("#formRegistro").on("submit", (event) => {
+    event.preventDefault();
+    var email = $("#email").val();
+    var password = $("#password").val();
+    if (emailCorrect == false && passwordCorrect == false) {
+        $("#registerText").fadeOut();
+        $("#infoText").fadeIn();
+    } else {
+
+        auth.createUserWithEmailAndPassword(email, password).then(userCredential => {
+            $("#registerText").text(email + " ha sido registrado");
+            $("#registerText").fadeIn();
+            setTimeout(() => {
+                $("#registerText").fadeOut();
+            }, 2000);
+            $("#infoText").fadeOut();
+            $("#formRegistro").trigger("reset");
+            emailCorrect = false;
+            passwordCorrect = false;
+        }).catch(error => {
+            var errorMessage = error.message;
+            $("#infoText").text(errorMessage);
+            $("#infoText").fadeIn();
+        });
+
+    }
+
+    $("#btnCancelar").click(() => {
+        $("#infoText").hide();
+    });
+});
+
+// Cerrando la sesión
+
+$("#logOut").click(() => {
+    auth.signOut().then(() => {
+        location.href = "admin.html";
+    });
+});
+
+
+
+
+//  Métodos de validación de campos
+
+$("#email").blur(() => {
+    var estado = checkEmail();
+    if (estado) {
+        $("#emailText").hide();
+        $("#email").removeClass("border-invalid");
+        emailCorrect = true;
+    } else {
+        $("#emailText").show();
+        $("#email").addClass("border-invalid");
+        emailCorrect = false;
+    }
+});
+
+$("#password").blur(() => {
+    var estado = checkPassword();
+    if (estado) {
+        $("#passwordText").hide();
+        $("#password").removeClass("border-invalid");
+        passwordCorrect = true;
+    } else {
+        $("#passwordText").show();
+        $("#password").addClass("border-invalid");
+        passwordCorrect = false;
+    }
+});
+
+function checkEmail() {
+    var pattern = /^([a-zA-Z0-9._%-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,6})*$/;
+    var inputEmail = $("#email").val();
+    var estado = true;
+
+    if (pattern.test(inputEmail) && inputEmail !== "") {
+        estado = true;
+    } else {
+        estado = false;
+    }
+
+    return estado;
+}
+
+function checkPassword() {
+    var pattern = /(?=(.*[0-9]))((?=.*[A-Za-z0-9])(?=.*[A-Z])(?=.*[a-z]))^.{8,}$/;
+    var inputPassword = $("#password").val();
+    var estado = true;
+
+    if (pattern.test(inputPassword) && inputPassword !== "") {
+        estado = true;
+    } else {
+        estado = false;
+    }
+
+    return estado;
 }
